[refactor/#63] reservation-history 페이지 컴포넌트 구조 개선
